[Org] Don’t make NOP inherited when looking for stuck projects
This is a major change in the semantics of this tag: NOP now does not designated
whole projects anymore, but only headlines that should not considered by
themselves as projects.  Consequently, sub-headlines of NOP-tagged headlines can
by projects on their own, and also show up as stuck projects if they are so.