[Org] Extend functionality to insert past interruptions
For this a new function ‘timeline-tools-clockline-no-org-agenda-conflicts’ as
been added that reads in a clock line (start and end times), updates all org
agenda files to fix conflicting clock lines, and returns the string of the new
clock line.  This works for all complete clock lines (i.e., those with end
times), but not for open clock lines (i.e., those missing an end timestamp).