[Utils] Fix directory name for ‘db/dired-from-shell-command’
Sometimes the string "Output of ‘…’" seems to be interpreted as directory name,
and then dired complains about it not being existent.  It is not completely
understood yet when this will happen, and an easy workaround seems to be to
stick to a fixed string that works.  This is what’s used now.