Make sure `message-mode' uses the right function to send mail
Up to now, it seemed to be sufficient to set `send-mail-function' alone, but
somehow some changes have made it necessary to set `message-send-mail-function'
explicitly.  If not done, it defaults to `message-send-mail-with-sendmail',
using the sendmail exectuable to send mail.  At least on my machine this results
in the message being delivered to the local exim instance, which does not allow
sending remote sending of mail.