Disable custom setter for special agenda files for now
It is no good to update `org-agenda-files' when setting those variables, only to
be overwritten by customize itself latter on.  Maybe it's better to instead have
a custom setter that updates the variable, but also checks whether the file is
also contained in `org-agenda-files', warning the user if this is not the case?