Update criteria for stuck projects
Previously, a project was not stuck if any subtask was tagged NOP (no-project).
However, this led constellations like the following to be hidden from the stuck
project list:

  * Test               :NOP:
  ** TestTest
  *** TestTestTest     :NOP:

In this case, the NOP at TestTestTest would result in hiding TestTest, which,
however, does not have any more things to do and should thus be marked stuck.
The new configuration will check NOP only at the top headline, and not at any
other sub-headlines.  For this, a property search TAGS={NOP} is necessary,
because otherwise tag inheritance would result in wrong false negatives.