Add buffer overview to main helm shortcut menu
Often, I want to open some file and try to first open the corresponding buffer
in the hope there's some open buffer for it already.  When no such buffer
exists, I have to close the list of buffers, reopen the helm shortcut menu, and
navigate to the corresponding bookmark.  This is cumbersome and somehow
duplicate work, since the bookmark and the file most often are named similarly.

With a buffer overview in the helm shortcut menu I can now search for the file
directly and if it's already open, directly select the corresponding buffer.
If it has not been opened, though, I can navigate to the corresponding bookmark
with at most one additional keystroke (C-o) and open it there.  Nice!