Mark task as not stuck while attending last meeting
If the only item below a task is a meeting that is currently attended, and thus
marked with the ATTN keyword, then the task as stuck.  Indeed, during the
meeting, new items may be added to the task at hand, so that after the meeting
is finished, the task will still be not stuck.  Because of this,
`org-stuck-projects' has been updated to not classify tasks as stuck if an item
with ATTN keyword is attached to it.