Introduct org capture template for complex tickets
From time to time, a new item introduces a more complex task that requires a
separate headline.  The new capture template is meant for this use case: it will
ask for a ticket number (the complex tasks I currently have to deal with always
come with a ticket number), a headline description, a reference (link to
redmine ticket or something), and a first task (usually a link to the first
email from the ticket).  It might be that the template is too specific, but
let's try it out first and adapt it if necessary.