Remove unused `hydra-shortcuts'
The idea of having a hydra to access frequently used features is certainly nice,
but quite hard to achive when one wants to redefine the hydra every time
`db/frequently-used-features' changes.  Regrettably, there are not „ephemeral
hydras“ that are created every time one would like to access it.  Therefore, the
shortcuts hydra is removed for now, but may come back again when we have found a
better way to handle its dynamic nature.