Rename main music hydra and move it to `db-music'
The `db-music' package is supposed to be an abstract interface to music
functionality, and should thus define the main hydra for this.  Moreover, the
hydra should not contain „emms“ in it's name, although it's using only EMMS
functions.

It's not clear whether EMMS will every be replaced by some other backend, but
it's nicer to have a (more or less) clear separation between user frontend and
implementation backend.