Restructure capture templates
The template for simple tasks should be simple, but got more complex in the last
change.  This mistake is reverted with this commit, and the usual "t" can now be
used again for simple tasks.  The "T" shortcut now provides capture templates
for complex tasks, but since complex tasks are both manifold and rarer, the
capture templates are provided with two-letter shortcuts.