Rework custom agenda views
The agendas for monthly reviews and notes are gone, I don't use them.  My
monthly review does not look at the agenda at all, only at my backlog.
Searching notes is done using Org's search functionality, and it's working very
fine.

The „Everything“ agenda has been renamed to „Unsupervised“.  The name
„Everything“ is not apt anymore, as this view clusters items that are not
displayed in the main agenda or are stuck.  The name „Unsupervised“ is a little
bit better, but may also change in the future.

Add a new view for listing all items that are not periodic and not part of the
personal backlog (i.e., not tagged with SOMEWHEN).  This list should help
getting an overview over the current workload of real TODO tasks.  Note that
both periodic tasks (tagged with PERIODIC, i.e., series of tasks) and regular
tasks (tagged with REGULAR, i.e., repeating tasks) are excluded from this list.

Descriptions of agenda views have been extended whereever it felt necessary.