Add custom agenda view for projects
In our configuration, projects are headlines without TODO keywords, that are
neither periodic tasks nor dates (dates usually have todo keywords, namely GOTO
or ATTN, but sometimes dates are just for information, and then the keyword is
missing).  Additionally, projects are not tagged with NOP.

This listing helps to see the current projects one is working on.