Eagerly load helm
Autoloading helm does not work well with custom keybindings.  Binding our
default "C-c h" to either `helm-command-map' or `helm-command-prefix' gives
errors, as both are not commands.  In the previous configuration, the prefix
"C-c h" was initially undefined and only defined when helm was loaded.  This led
to irritating behavior.

All this can be fixed by eagerly loading helm.  This may slow down startup, in
particular on Windows, but it should be worth it.