Update key bindings for handling Org Mode links
F9 is a comparably prominent key binding, and we now bind it to the more
important `db/org-add-link-to-other-item'.  The formar binding to
`db/org-find-links-to-current-item' is bound to F11 now, since it will still be
used often, but not more often then inserting links (I think).  The old binding
for F11 to `org-capture' has not been used much, and so we can dispense of it.