Do not partially fetch articles
This leads to errors when forgetting to fetch the complete article before
forwarding it.  Moreover, redisplaying the article sometimes leads to a
completely fetched article to be displayed only partially again, resulting in
repeated fetched.  Finally, the performance gain is not worth the effort in my
setup, so let's just disable it.