Do not use ivy for completion in regions
This makes eshell completion use drop down menues instead of the standard
complete-until-ambiguous style I am used to.  Moreover, in Org Babel shell
source blocks, ivy completion in region causes a drop down menu of possible
commands to appear after a block has been created and entered for the first
time, which not only is annoying, but also sometimes causes to cursor to
disappear (set `cursor-type' to 'bar to revert).