Disable native handling of tabs in org source blocks
When company is active, this leads to completions offered as a first step in
shell blocks, for example, which is both unnecessary and triggers a bug in
ivy (the cursor is invisible afterwards).  Furthermore, this completion is
almost always unnecessary, as the snippets a usually either quite simple or
copied from elsewhere.  If completion is still required, editing the code block
within the native major mode (C-c ') should be sufficient.