Disable candidate restriction in ivy
The main motivation is me often mistyping S-SPC for SPC, loosing my current
candidates and frustrating myself.  This is even more frustrating when inserting
text into non-selection inputs (org-capture), which is why we had disabled ivy
completion for org-capture previously.  Since we are no using
ivy--regex-ignore-order for candidate regex building anyway, input restriction
is not necessary anymore, and so we remove the shortcut altogether.