Fix a bug when checking signatures accidentally containing CRLF
From our previous attempts to get S/MIME working when using Outlook, the
function `mm-copy-to-buffer` had been adviced to change all LF to CRLF, to allow
`mm-copy-to-buffer` to find the beginning of the body.  However, when the body
itself is binary and accidentally contains an CRLF, that content is modified as
well.  This might break the content, as it happend with the signature of
elpa.gnu.org, which indeed contained an CRLF.

We are now a bit more non-invasive and replace the original version of
`mm-copy-to-buffer` with one that explicitly also searches for ^\r\n in addition
to just ^\n.  Let's see how good that will work.  Signature checking for
elpa.gnu.org is functioning again, and S/MIME still seems to be working.