Do not sum up efforts in Org column view
Indeed, in my workflow, the effort estimate of an item is independent of the
effort estimates of all its subitems.  It thus does not makes to sum them up,
and indeed the "{:}" in the column view format causes the (independent) effort
estimate to be overwritten by the sum of the efforts of its subitems every time
column view is turned on.