Explicitly set TRAMP default method
Using sshx as default on non-Windows systems seems to be a more robust choice
than using ssh, because the latter might be prone to errors caused by my
user-specific shell configuration.  And while we are at it, using plink on
Windows is more reasonable than using ssh, at least in my working environment.