Never insert terminating space when completing in eshell
eshell is using pcomplete as completion meta-framework, which by default will
insert the value of `pcomplete-termination-string` to each finished completion.
However, when using newer versions of `ivy` for path completion, each completed
directory within a path is considered a finished match, and pcomplete will
insert a blank.  This is annoying, and since we do not require the final blank
otherwise, we can equally just disable it.