Use only file name in Org mode refile targets instead of buffer name
Using the buffer name does not seem to play well together with org refile
keeping the last choice on top of the candidate list, because a slash is
automatically appended to this entry.  This results in an invalid refile
location.  Thus reverting back to only using the file name, it's fine this way.