Fix Org capture templates for Read and Response
The Read capture template did not properly ask for the topic and did not have a
placement indicator for the cursor, fixed that.  The Response capture template
now has a “Reply ” action word to indicate the proper action, and also has a
placement indicator in it.  This way, it should be easier to use these capture
templates.