Sort items in Waiting-For list by priority and category
This is the same sorting as it is used in the Next Action list, and should
provide a better overview over the current items on the Waiting-For list.

Formatted the value for `org-agenda-sorting-strategy` on one line for better
readability, and also changed that for the definition of the Next Action list.