Move refile overview to unspecified agenda
Things to refile should not be shown on the main agenda, as this view is only
meant to show open tasks.  Indeed, if to-be-refiled items are already done, they
should not be here for exactly that reason; and if they are not done yet, then
they will show up in the Next Action list anyway.

Bringing items to their (supposedly) correct location is not part of doing the
item, it's part of cleaning up (e.g., while doing the weekly review).

Emptying the refile file should still be done regularly, though, i.e., every
day (via a daily review) or weekly (in the weekly review).