Do not keep inserted links in history
When keeping old links around, it may happen that adding another link is already
present in `org-stored-links`.  In that case, only a message is displayed (“This
link already exists”), but the link itself is not pushed to the front of the
link history.  A subsequent insertion of the link then requires manual selection
of the desired target, resulting in massive interruptions of the current flow.