Disable completion in Org Mode buffers
I do not use the completion provided by company-mode, and sometimes it's even
disruptive during writing.  So let's disable it for the time being.

The original reason to keep it activated has been to have completion when
inserting links to local files.  This can better be achieved by providing a
universal argument to `org-insert-link`.