Removed capture templates that add stuff to the current clock
Those things are added at the end of the LOGBOOK, instead of at the beginning.
This does not help, I would rather like to have those things being added as if I
would have taken a note on the currently clocked-in task.  Indeed, this is what
I have been doing instead of using those capture template, which is why this
commit removes them.