Allow to ignore stuck projects that are scheduled in the future
This allows to postpone projects into the future when they are not relevant now
but would otherwise be stuck.  Because of the scheduling, an automatic reminder
will appear on the agenda when the date is due, upon which the projects is shown
again as stuck.  Then new items can be planned, or other measure might be taken.