Do not show scheduled items in WIP or NEXT list
When things are scheduled, they are shown in the time grid portion of the
agenda.  When they are scheduled in the future, they will not show up until that
date is due; when they are scheduled in the past or present, they are shown on
the time grid directly.  Both situations are sufficient, and it's thus not
necessary to show scheduled items in other lists as well.