Add org-ql as extended query languag
This includes, among others, a dynamic block to insert the result of a query –
which is exactly what I am looking for.  I have to learn a new query language,
though, but it seems as if non-sexp syntax is easy enough; and having a
lisp-like query syntax is undoubtedly a big improvement!