Disable lispy in minibuffer
Somehow, using lispy in the minibuffer causes random indentations although
everythin is on the same line.  This might not be an issue with lispy itself,
disabling it however makes the symptom go away.  Since I am not using lispy
features in the minibuffer anyway, I can as well switch it off.