Ignore active future appointments on the WIP list
When appoints have been started but are then postponed to the future, I do not
want to see them on the WIP list.  I could schedule them for the new date in
addition, but this would count the associated effort twice in the agenda view.
So let's just ignore timestamps in the future, they will show up when they are
due anyway.