Update way scheduled items are ignored in custom Org agendas
Items scheduled now or in the future are now ignored by default.  Rationale: if
things are scheduled now or in the future, they are displayed on the main agenda
and not overdue, so there's no need to show them in the other agendas (meant for
reviewing); if an item is scheduled in the past, it's still shown on the main
agenda, but should also be reviewd for why it's late, so we include it in the
other agendas as well.

Also ignore NOTE items that are scheduled, as the same logic applies there as
well.