Include base directory of current git repository in eshell prompt
This allows to see of which git repository the current working directory is part
of.

This is also included in Howard Abrams configuration, but he instead determines
the name of the current git repository by looking at `remote.origin.url`.
However, my upstream branch is not always called `origin`, and sometimes there
is no upstream repository at all.  Using the name of the current directory (not
the whole path, though) instead seems to be a good compromise from my point of
view.