Remove redundant agenda prefix definitions
These definitions are easily covered by the global setting of
`org-agenda-prefix-format` and can thus also be customized now.  Note that this
yields some extra space with the default setting, as the entries of the custom
project agenda usually do not have an effort property set – but I think this is
easily bearable.