Replace personal function redefinitions with around advices
This way the original function definition is left intact and can be found via
the help buffer.

The manual overwrite of `org-ql--link-regexp` has been left as it is, as it's a
`cl-defun` definition and I am not quite sure how turn this into an ELisp around
advice without breaking things.

The manual fix for `enriched-decode-display-prop` has also been left untouched,
maybe this should be an around advice as well?