Fix `org-read-date` not recognizing hh:mm parts on startup
The problem seems to be the dynamically scoped variable `org-time-was-given`,
which is used by `org-read-date` to decided whether an hh:mm part is present.
The variable `org-time-was-given` is set by `org-read-date-analyze`, but only
when it's (globally) bound, which it is not on startup (since it's only declared
via `defvar`).  Manually setting the variable to nil binds the variable and
everything seems to work nicely.