Reverse treatment of universal argument for helm shortcut menu
Usually, when searching for important documents vie `db/helm-shortcuts`, I
forget to include the universal argument, abort the current search and repeat
it.  This is annoying.  Since listing documents is fast enough, including them
by default should alleviate this issue.

When retrieving the list of documents is too slow and should be disabled
interactively, providing a universal argument excludes this listing now.